backend/pyproject.toml
[tool.poetry]
name = "agentic-model-training-platform"
version = "0.1.0"
description = "Autonomous AI system for machine learning model development"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
celery = "^5.3.6"
redis = "^5.0.1"
google-cloud-aiplatform = "^1.42.1"
google-generativeai = "^0.3.2"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
google-cloud-storage = "^2.14.0"
google-cloud-logging = "^3.9.0"
websockets = "^12.0"
httpx = "^0.26.0"
pandas = "^2.1.4"
numpy = "^1.26.3"
scikit-learn = "^1.4.0"
pillow = "^10.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^23.12.1"
ruff = "^0.1.11"
mypy = "^1.8.0"
locust = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
